feat: File restructure
