+++ components/AllFlightsScreen.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { View, Text } from 'react-native'
+
+function AllFlightsScreen() {
+  return (
+    <View>
+      <Text>Previous Flights</Text>
+      {/* Add your previous flights component here */}
+    </View>
+  )
+}
+
+export default AllFlightsScreen
