+++ components/LogFlightScreen.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { View, Text } from 'react-native'
+
+function LogFlightScreen() {
+  return (
+    <View>
+      <Text>Log Flight</Text>
+      {/* Add your log flight component here */}
+    </View>
+  )
+}
+
+export default LogFlightScreen
